% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiEstim_wrap.R
\name{EpiEstim_wrap}
\alias{EpiEstim_wrap}
\title{Hello World}
\usage{
EpiEstim_wrap(
  I_incid,
  si_distr,
  t_window,
  overlap,
  mean_prior = 5,
  std_prior = 5
)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This is an example of how to create and document exported functions.
}
