% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_Rt_wrap.R
\name{gam_Rt_wrap}
\alias{gam_Rt_wrap}
\title{Wrapper to run the gam version of EpiEstim with temporal smooth}
\usage{
gam_Rt_wrap(I_incid, si_distr, dist = "poisson", k_basis = 10)
}
\arguments{
\item{I_incid}{A dataframe of non-negative integers with two columns, so that
\code{I_incid$local} contains the incidence of cases due to local transmission
and \code{I_incid$imported} contains the incidence of imported cases (with
\code{I_incid$local + I_incid$imported} the total incidence).}

\item{si_distr}{a vector containing
the discrete serial interval distribution(s) used for estimation. 
This is equivalent to the input in EpiEstim when using the "non_parametric_si"
method.}

\item{dist}{distribution assumed for gam, currently support 'poisson' (default) or 'nb' 
for negative binomial (see mgcv::nb(link = "log")).}
}
\value{
A list including all results, including:

model: the output of the gam model (equivalent to the output when running mgcv::gam);

Rt: a dataframe, including the time step for estimation; the mean, lower and higher 
             quantile (95% CI), and standard deviation in the estimated Rt at each time 
             step (center of each time-window).
}
\description{
Run the gam equivalent of EpiEstim.with temporal smoothing.
}
