% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_rt.R
\name{kernel_rt}
\alias{kernel_rt}
\title{Kernel for simulated Rt}
\usage{
kernel_rt(map, mu, sigma, Rt_min_max)
}
\arguments{
\item{map}{a dataframe (usually will be a GIS object) with one row per location and
containing centroid coordinates of each location. Centroid coordinates should be named
map$cent.x and map$cent.y.}

\item{mu}{a list which size correspond to the number of simuated normal kernels. Each element
of the list is a 2D vector containing the coordinates (x,y) of 
the central position of each normal kernel .}

\item{sigma}{a vector of the same size as mu containing the variances
for each of the simulated normal kernels.}

\item{Rt_min_max}{a dataframe containing Rt_min_max$min and Rt_min_max$min, as the minimum and maximum
simulated Rts.}
}
\value{
A vector R of simulated Rts with one Rt per location.
}
\description{
This function, used in the proof of concept paper, allow the user to obtain spatially smoothed
simulated Rts. As described in the paper methods the Rts are simulated from multiple bivariate normal distributions
that are scaled to a given min and max.
}
