% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare2Rt.R
\name{plot_compare2Rt}
\alias{plot_compare2Rt}
\title{Function for plotting}
\usage{
plot_compare2Rt(a, b, t_window, overlap, Corr = FALSE, threshold = NA)
}
\description{
This function is for internal use only and is not documented.
It was used to compare fit of multiple methods in the proof-of-concept paper.
}
