% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_Rtglm_location.R
\name{pred_Rtglm_location}
\alias{pred_Rtglm_location}
\title{Output the predictions for Rt}
\usage{
pred_Rtglm_location(model, newdata, t_window = NA, overlap = FALSE, location)
}
\arguments{
\item{newdata}{A dataframe output of the Rt.glm function 'prep_glm_tWindow'. It includes the time (t), 
incidence (incidence), the overall infectivity (Oi), its log-transform (log_Oi) and the time-windows (tw).}

\item{t_window}{a single integer characterising the time-window to be used. During
a time-window, the Rt is assumed to be constant. Time-windows may be overlapping 
or not (see overlap parameter)}

\item{overlap}{a logical, TRUE or FALSE, indicating whether time-window used for estimation
should overlap. by default, overlap is set at FALSE.}

\item{model:}{the output of the glm model (equivalent to the output when running mgcv::gam);}
}
\value{
newdata: a dataframe, same as input but with added estimates: the mean, lower and higher 
             quantile (95% CI), and standard deviation in the estimated Rt at each time 
             step (center of each time-window when overlapping).
}
\description{
This functions translate the results of the glm back into Rt estimations.
}
