% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_glm.R
\name{prep_glm}
\alias{prep_glm}
\title{Hello World}
\usage{
prep_glm(I_incid, si_distr)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This is an example of how to create and document exported functions.
}
