% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_glm_tWindow.R
\name{prep_glm_tWindow}
\alias{prep_glm_tWindow}
\title{Hello World}
\usage{
prep_glm_tWindow(data_infer, t_window, overlap)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This is an example of how to create and document exported functions.
}
