% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiEstim_sp_wrap.R
\name{EpiEstim_sp_wrap}
\alias{EpiEstim_sp_wrap}
\title{Wrapper to run Epiestim}
\usage{
EpiEstim_sp_wrap(I_incid, si_distr, t_ini, mean_prior = 5, std_prior = 5)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This function is for internal use only and is not documented.
It was used to run EpiEstim with fixed parameterisation in the proof-of-concept paper, and
obtain standardised outputs.
}
