% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiEstim_wrap.R
\name{EpiEstim_wrap}
\alias{EpiEstim_wrap}
\title{Wrapper to run Epiestim}
\usage{
EpiEstim_wrap(
  I_incid,
  si_distr,
  t_window,
  overlap,
  mean_prior = 5,
  std_prior = 5
)
}
\description{
This function is for internal use only and is not documented.
It was used to run EpiEstim with fixed parameterisation in the proof-of-concept paper, and
obtain standardised outputs.
}
