% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_Rt_wrap.R
\name{gam_Rt_wrap}
\alias{gam_Rt_wrap}
\title{Hello World}
\usage{
gam_Rt_wrap(I_incid, si_distr, t_window, overlap)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This is an example of how to create and document exported functions.
}
