% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_Rt_wrap_corrTC.R,
%   R/glm_Rt_wrap_corrTC_location.R
\name{glm_Rt_wrap_corrTC}
\alias{glm_Rt_wrap_corrTC}
\title{Wrapper to run the glm version of EpiEstim with correction}
\usage{
glm_Rt_wrap_corrTC(
  I_incid,
  rho,
  gamma,
  location,
  si_distr,
  t_window,
  overlap = FALSE
)

glm_Rt_wrap_corrTC(
  I_incid,
  rho,
  gamma,
  location,
  si_distr,
  t_window,
  overlap = FALSE
)
}
\arguments{
\item{I_incid}{a list of dataframe of non-negative integers with two columns, so that
\code{I_incid$local} contains the incidence of cases due to local transmission
and \code{I_incid$imported} contains the incidence of imported cases (with
\code{I_incid$local + I_incid$imported} the total incidence).
there is one dataframe per location.}

\item{rho}{a matrix containing
the proportion of T/S over time, should have the same number of row as I_incdid, and column as number of locations}

\item{gamma}{to account for censoring, probability that a case (S or T or T+) is already in the database.
As for rho, should be a matrix with the same number of row as I_incdid, and column as number of locations}

\item{location}{a vector of location names, the length should be the same as number of columns in rho/gamma and the dataframes in the I_incid list}

\item{si_distr}{a vector containing
the discrete serial interval distribution(s) used for estimation. 
This is equivalent to the input in EpiEstim when using the "non_parametric_si"
method.}

\item{t_window}{a single integer characterising the time-window to be used. During
a time-window, the Rt is assumed to be constant. Time-windows may be overlapping 
or not (see overlap parameter)}

\item{overlap}{a logical, TRUE or FALSE, indicating whether time-window used for estimation
should overlap. by default, overlap is set at FALSE.}
}
\value{
A list including all results, including:

config: a record of the configuration, i.e. a list including the parameters
             t_window and overlap

model: the output of the glm model (equivalent to the output when running mgcv::gam);

Rt: a dataframe, including the time step for estimation; the mean, lower and higher 
             quantile (95% CI), and standard deviation in the estimated Rt at each time 
             step (center of each time-window).

A list including all results, including:

config: a record of the configuration, i.e. a list including the parameters
             t_window and overlap

model: the output of the glm model (equivalent to the output when running mgcv::gam);

Rt: a dataframe, including the time step for estimation; the mean, lower and higher 
             quantile (95% CI), and standard deviation in the estimated Rt at each time 
             step (center of each time-window).
}
\description{
Run the glm equivalent of EpiEstim, with correction for testing capacity and censoring

Run the glm equivalent of EpiEstim, with correction for testing capacity and censoring for multiple locations
generally same as the glm_Rt_wrap function, but with some addition
}
