% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_rt.R
\name{kernel_rt}
\alias{kernel_rt}
\title{Hello World}
\usage{
kernel_rt(uk, mu, sigma, Rt_min_max)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This is an example of how to create and document exported functions.
}
