% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare2Rt.R
\name{plot_compare2Rt}
\alias{plot_compare2Rt}
\title{Hello World}
\usage{
plot_compare2Rt(a, b, t_window, overlap, Corr = FALSE, threshold = NA)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This is an example of how to create and document exported functions.
}
