% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_Rtglm.R
\name{pred_Rtglm}
\alias{pred_Rtglm}
\title{Hello World}
\usage{
pred_Rtglm(model, newdata, overlap = FALSE)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This is an example of how to create and document exported functions.
}
