% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_glm_sp.R
\name{prep_glm_sp}
\alias{prep_glm_sp}
\title{Hello World}
\usage{
prep_glm_sp(I_incid, si_distr, x, y)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This is an example of how to create and document exported functions.
}
